+++ src/core/client/api/consts/cameraConst.ts
@@ -1,4 +1,3 @@
-import { CameraTarget } from '@AthenaClient/systems/cameraTarget';
 import { GameplayCamera } from '@AthenaClient/systems/gameplayCamera';
 import PedEditCamera from '@AthenaClient/utility/camera';
 import { CinematicCam } from '@AthenaClient/utility/cinematic';
@@ -7,5 +6,4 @@ export const cameraConst = {
     ped: PedEditCamera,
     gameplay: GameplayCamera,
     cinematic: CinematicCam,
-    target: CameraTarget,
 };