+++ src/core/client/api/consts/wheelMenus.ts
@@ -1,5 +1,4 @@
 import { NpcWheelMenu } from '@AthenaClient/menus/npc';
-import { ObjectWheelMenu } from '@AthenaClient/menus/object';
 import { PlayerWheelMenu } from '@AthenaClient/menus/player';
 import { VehicleWheelMenu } from '@AthenaClient/menus/vehicle';
 
@@ -9,9 +8,6 @@ export const wheelMenusConst = {
     },
     player: {
         add: PlayerWheelMenu.addInjection,
-    },
-    object: {
-        add: ObjectWheelMenu.addInjection,
     },
     vehicle: {
         add: VehicleWheelMenu.addInjection,